Complete Walk Rating
Complete walk now takes in a rating which would be given by the owner of the walk.

ids are important when it comes to incrementing and adjusting the rating of a walker.

Home page now displays the related walk miles, completed and processing.

Profile correctly displays miles walked.

Completing a walk now requests a rating for the walker.